requirements.txt
Django>=5.0,<5.1
gunicorn>=21.2.0
psycopg2-binary>=2.9.9
redis>=5.0.1
groq>=0.4.0
python-dotenv>=1.0.0
Pillow>=10.1.0
whitenoise>=6.6.0
django-cors-headers>=4.3.1
PyPDF2>=3.0.1
requests>=2.31.0
google-auth>=2.22.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.95.0
django-allauth>=0.57.0
